Move test class to more appropriate place
